feeds table cleanup and make column names consistent
