+++ pdr_backend/pdr_dashboard/pages/feeds.py
@@ -1,23 +1,17 @@
-from typing import List, Dict, Any, Tuple, Union
-
 from dash import html, dcc, dash_table
-from pdr_backend.pdr_dashboard.util.data import (
-    get_feed_column_ids,
-    find_with_key_value,
-)
 from pdr_backend.pdr_dashboard.dash_components.view_elements import (
     get_metric,
     get_search_bar,
 )
-from pdr_backend.pdr_dashboard.util.format import format_table
+from pdr_backend.pdr_dashboard.util.data import get_feeds_data_for_feeds_table
 from pdr_backend.pdr_dashboard.pages.common import TabularPage, Filter, add_to_filter
 from pdr_backend.pdr_dashboard.dash_components.modal import get_modal
 
 filters = [
     {"name": "base_token", "placeholder": "Base Token", "options": []},
     {"name": "quote_token", "placeholder": "Quote Token", "options": []},
-    {"name": "exchange", "placeholder": "Exchange", "options": []},
-    {"name": "time", "placeholder": "Time", "options": []},
+    {"name": "source", "placeholder": "Source", "options": []},
+    {"name": "timeframe", "placeholder": "Timeframe", "options": []},
 ]
 
 filters_objects = [Filter(**item) for item in filters]
@@ -120,8 +114,8 @@ class FeedsPage(TabularPage):
             self.app.network_name
         )
 
-        feed_cols, feed_data, raw_feed_data = self.get_feeds_data_for_feeds_table(
-            feed_stats, feed_subscriptions
+        feed_cols, feed_data, raw_feed_data = get_feeds_data_for_feeds_table(
+            self.app.feeds_data, feed_stats, feed_subscriptions
         )
 
         self.app.feeds_table_data = raw_feed_data
@@ -153,96 +147,3 @@ class FeedsPage(TabularPage):
             ],
             style={"width": "100%", "overflow": "scroll"},
         )
-
-    def get_feeds_stat_with_contract(
-        self, contract: str, feed_stats: List[Dict[str, Any]]
-    ) -> Dict[str, Union[float, int]]:
-        result = find_with_key_value(feed_stats, "contract", contract)
-
-        if result:
-            return {
-                "avg_accuracy": float(result["avg_accuracy"]),
-                "avg_stake_per_epoch_(OCEAN)": float(result["avg_stake"]),
-                "volume_(OCEAN)": float(result["volume"]),
-            }
-
-        return {
-            "avg_accuracy": 0,
-            "avg_stake_per_epoch_(OCEAN)": 0,
-            "volume_(OCEAN)": 0,
-        }
-
-    def get_feeds_subscription_stat_with_contract(
-        self, contract: str, feed_subcription_stats: List[Dict[str, Any]]
-    ) -> Dict[str, Union[float, int, str]]:
-        result = find_with_key_value(feed_subcription_stats, "contract", contract)
-
-        if result:
-            sales_str = f"{result['sales']}"
-
-            df_buy_count_str = (
-                f"{result['df_buy_count']}-DF" if result["df_buy_count"] > 0 else ""
-            )
-            ws_buy_count_str = (
-                f"{result['ws_buy_count']}-WS" if result["ws_buy_count"] > 0 else ""
-            )
-
-            if df_buy_count_str or ws_buy_count_str:
-                counts_str = "_".join(
-                    filter(None, [df_buy_count_str, ws_buy_count_str])
-                )
-
-                if counts_str:
-                    sales_str += f"_{counts_str}"
-
-            return {
-                "price_(OCEAN)": result["price"],
-                "sales": sales_str,
-                "sales_raw": result["sales"],
-                "sales_revenue_(OCEAN)": result["sales_revenue"],
-            }
-
-        return {
-            "price_(OCEAN)": 0,
-            "sales": 0,
-            "sales_raw": 0,
-            "sales_revenue_(OCEAN)": 0,
-        }
-
-    def get_feeds_data_for_feeds_table(
-        self,
-        feed_stats: List[Dict[str, Any]],
-        feed_subcription_stats: List[Dict[str, Any]],
-    ) -> Tuple[List[Dict[str, str]], List[Dict[str, Any]], List[Dict[str, Any]]]:
-
-        temp_data = self.app.feeds_data
-
-        new_feed_data = []
-        ## split the pair column into two columns
-        for feed in temp_data:
-            split_pair = feed["pair"].split("/")
-            feed_item = {}
-            feed_item["addr"] = feed["contract"]
-            feed_item["base_token"] = split_pair[0]
-            feed_item["quote_token"] = split_pair[1]
-            feed_item["exchange"] = feed["source"].capitalize()
-            feed_item["time"] = feed["timeframe"]
-            feed_item["full_addr"] = feed["contract"]
-
-            result = self.get_feeds_stat_with_contract(feed["contract"], feed_stats)
-
-            feed_item.update(result)
-
-            subscription_result = self.get_feeds_subscription_stat_with_contract(
-                feed["contract"], feed_subcription_stats
-            )
-
-            feed_item.update(subscription_result)
-
-            new_feed_data.append(feed_item)
-
-        columns = get_feed_column_ids(new_feed_data[0])
-
-        formatted_data = format_table(new_feed_data, columns)
-
-        return columns, formatted_data, new_feed_data
