+++ pdr_backend/pdr_dashboard/test/json_fixtures/expected_feeds_table_data.json
@@ -4,8 +4,8 @@
     "Addr": "0xf28...70b30",
     "Base Token": "TRX",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "65.09%",
     "Avg Stake Per Epoch (Ocean)": "193.28",
     "Volume (Ocean)": "40.98K",
@@ -18,8 +18,8 @@
     "Addr": "0xd41...90065",
     "Base Token": "BNB",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "975.06",
     "Volume (Ocean)": "15.6K",
@@ -32,8 +32,8 @@
     "Addr": "0x2d8...a3152",
     "Base Token": "XRP",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "1.02K",
     "Volume (Ocean)": "16.27K",
@@ -46,8 +46,8 @@
     "Addr": "0x3fb...6f3bc",
     "Base Token": "SOL",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "49.06%",
     "Avg Stake Per Epoch (Ocean)": "182.83",
     "Volume (Ocean)": "38.76K",
@@ -60,8 +60,8 @@
     "Addr": "0x74a...c06dd",
     "Base Token": "SOL",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "948.45",
     "Volume (Ocean)": "15.18K",
@@ -74,8 +74,8 @@
     "Addr": "0xfa6...b08dd",
     "Base Token": "LTC",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "998.29",
     "Volume (Ocean)": "15.97K",
@@ -88,8 +88,8 @@
     "Addr": "0x31f...ebfbd",
     "Base Token": "BNB",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "51.89%",
     "Avg Stake Per Epoch (Ocean)": "207.1",
     "Volume (Ocean)": "43.91K",
@@ -102,8 +102,8 @@
     "Addr": "0xbe0...2d434",
     "Base Token": "DOGE",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "48.58%",
     "Avg Stake Per Epoch (Ocean)": "187.3",
     "Volume (Ocean)": "39.71K",
@@ -116,8 +116,8 @@
     "Addr": "0x816...b2553",
     "Base Token": "BTC",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "870.51",
     "Volume (Ocean)": "13.93K",
@@ -130,8 +130,8 @@
     "Addr": "0xd49...e0a80",
     "Base Token": "TRX",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "994.48",
     "Volume (Ocean)": "15.91K",
@@ -144,8 +144,8 @@
     "Addr": "0x30f...77695",
     "Base Token": "ETH",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "48.94%",
     "Avg Stake Per Epoch (Ocean)": "159.64",
     "Volume (Ocean)": "37.51K",
@@ -158,8 +158,8 @@
     "Addr": "0x9c4...07e1f",
     "Base Token": "DOT",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "52.36%",
     "Avg Stake Per Epoch (Ocean)": "195.97",
     "Volume (Ocean)": "41.54K",
@@ -172,8 +172,8 @@
     "Addr": "0xb1c...23766",
     "Base Token": "LTC",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "49.06%",
     "Avg Stake Per Epoch (Ocean)": "192.51",
     "Volume (Ocean)": "40.81K",
@@ -186,8 +186,8 @@
     "Addr": "0x18f...e3151",
     "Base Token": "ADA",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "50.47%",
     "Avg Stake Per Epoch (Ocean)": "207.16",
     "Volume (Ocean)": "43.92K",
@@ -200,8 +200,8 @@
     "Addr": "0x93f...8b48f",
     "Base Token": "DOT",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "982.41",
     "Volume (Ocean)": "15.72K",
@@ -214,8 +214,8 @@
     "Addr": "0xf8c...c3fb3",
     "Base Token": "DOGE",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "951.39",
     "Volume (Ocean)": "15.22K",
@@ -228,8 +228,8 @@
     "Addr": "0xe66...69690",
     "Base Token": "BTC",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "48.09%",
     "Avg Stake Per Epoch (Ocean)": "168.41",
     "Volume (Ocean)": "39.58K",
@@ -242,8 +242,8 @@
     "Addr": "0xa2d...a0cf0",
     "Base Token": "ADA",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "1.03K",
     "Volume (Ocean)": "16.49K",
@@ -256,8 +256,8 @@
     "Addr": "0x55c...2cedb",
     "Base Token": "XRP",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "49.79%",
     "Avg Stake Per Epoch (Ocean)": "143.39",
     "Volume (Ocean)": "33.7K",
@@ -270,8 +270,8 @@
     "Addr": "0xaa6...710c5",
     "Base Token": "ETH",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "931.99",
     "Volume (Ocean)": "14.91K",
