+++ pdr_backend/pdr_dashboard/test/json_fixtures/filtered_accuracy_min_55.json
@@ -4,8 +4,8 @@
     "Addr": "0xf28...70b30",
     "Base Token": "TRX",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "65.09%",
     "Avg Stake Per Epoch (Ocean)": "193.28",
     "Volume (Ocean)": "40.98K",
