+++ pdr_backend/pdr_dashboard/test/json_fixtures/filtered_base_token_eth_5m.json
@@ -4,8 +4,8 @@
     "Addr": "0x30f...77695",
     "Base Token": "ETH",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "48.94%",
     "Avg Stake Per Epoch (Ocean)": "159.64",
     "Volume (Ocean)": "37.51K",
