+++ pdr_backend/pdr_dashboard/test/json_fixtures/search_6f3bc.json
@@ -4,8 +4,8 @@
     "Addr": "0x3fb...6f3bc",
     "Base Token": "SOL",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "49.06%",
     "Avg Stake Per Epoch (Ocean)": "182.83",
     "Volume (Ocean)": "38.76K",
