+++ pdr_backend/pdr_dashboard/test/json_fixtures/search_ada.json
@@ -4,8 +4,8 @@
     "Addr": "0xa2d...a0cf0",
     "Base Token": "ADA",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "1.03K",
     "Volume (Ocean)": "16.49K",
@@ -18,8 +18,8 @@
     "Addr": "0x18f...e3151",
     "Base Token": "ADA",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "50.47%",
     "Avg Stake Per Epoch (Ocean)": "207.16",
     "Volume (Ocean)": "43.92K",
