+++ pdr_backend/pdr_dashboard/test/json_fixtures/search_eth.json
@@ -4,8 +4,8 @@
     "Addr": "0xaa6...710c5",
     "Base Token": "ETH",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "1h",
+    "Source": "Binance",
+    "Timeframe": "1h",
     "Avg Accuracy": "50.0%",
     "Avg Stake Per Epoch (Ocean)": "931.99",
     "Volume (Ocean)": "14.91K",
@@ -18,8 +18,8 @@
     "Addr": "0x30f...77695",
     "Base Token": "ETH",
     "Quote Token": "USDT",
-    "Exchange": "Binance",
-    "Time": "5m",
+    "Source": "Binance",
+    "Timeframe": "5m",
     "Avg Accuracy": "48.94%",
     "Avg Stake Per Epoch (Ocean)": "159.64",
     "Volume (Ocean)": "37.51K",
