+++ pdr_backend/pdr_dashboard/util/data.py
@@ -1,11 +1,14 @@
 import logging
-
 from copy import deepcopy
 from datetime import datetime, timedelta
-from typing import Union, List, Dict, Any, Optional
-from enforce_typing import enforce_types
+from typing import Union, List, Dict, Any, Optional, Tuple
 import dash
-from pdr_backend.pdr_dashboard.util.format import pick_from_dict, format_dict
+from enforce_typing import enforce_types
+from pdr_backend.pdr_dashboard.util.format import (
+    format_table,
+    pick_from_dict,
+    format_dict,
+)
 
 logger = logging.getLogger("predictoor_dashboard_utils")
 
@@ -128,3 +131,95 @@ def find_with_key_value(
             return obj
 
     return None
+
+
+def get_feeds_stat_with_contract(
+    contract: str, feed_stats: List[Dict[str, Any]]
+) -> Dict[str, Union[float, int]]:
+    result = find_with_key_value(feed_stats, "contract", contract)
+
+    if result:
+        return {
+            "avg_accuracy": float(result["avg_accuracy"]),
+            "avg_stake_per_epoch_(OCEAN)": float(result["avg_stake"]),
+            "volume_(OCEAN)": float(result["volume"]),
+        }
+
+    return {
+        "avg_accuracy": 0,
+        "avg_stake_per_epoch_(OCEAN)": 0,
+        "volume_(OCEAN)": 0,
+    }
+
+
+def get_feeds_subscription_stat_with_contract(
+    contract: str, feed_subcription_stats: List[Dict[str, Any]]
+) -> Dict[str, Union[float, int, str]]:
+    result = find_with_key_value(feed_subcription_stats, "contract", contract)
+
+    if result:
+        sales_str = f"{result['sales']}"
+
+        df_buy_count_str = (
+            f"{result['df_buy_count']}-DF" if result["df_buy_count"] > 0 else ""
+        )
+        ws_buy_count_str = (
+            f"{result['ws_buy_count']}-WS" if result["ws_buy_count"] > 0 else ""
+        )
+
+        if df_buy_count_str or ws_buy_count_str:
+            counts_str = "_".join(filter(None, [df_buy_count_str, ws_buy_count_str]))
+
+            if counts_str:
+                sales_str += f"_{counts_str}"
+
+        return {
+            "price_(OCEAN)": result["price"],
+            "sales": sales_str,
+            "sales_raw": result["sales"],
+            "sales_revenue_(OCEAN)": result["sales_revenue"],
+        }
+
+    return {
+        "price_(OCEAN)": 0,
+        "sales": 0,
+        "sales_raw": 0,
+        "sales_revenue_(OCEAN)": 0,
+    }
+
+
+def get_feeds_data_for_feeds_table(
+    feeds_data: List[Dict[str, Any]],
+    feed_stats: List[Dict[str, Any]],
+    feed_subcription_stats: List[Dict[str, Any]],
+) -> Tuple[List[Dict[str, str]], List[Dict[str, Any]], List[Dict[str, Any]]]:
+
+    new_feed_data = []
+    ## split the pair column into two columns
+    for feed in feeds_data:
+        split_pair = feed["pair"].split("/")
+        feed_item = {}
+        feed_item["addr"] = feed["contract"]
+        feed_item["base_token"] = split_pair[0]
+        feed_item["quote_token"] = split_pair[1]
+        feed_item["source"] = feed["source"].capitalize()
+        feed_item["timeframe"] = feed["timeframe"]
+        feed_item["full_addr"] = feed["contract"]
+
+        result = get_feeds_stat_with_contract(feed["contract"], feed_stats)
+
+        feed_item.update(result)
+
+        subscription_result = get_feeds_subscription_stat_with_contract(
+            feed["contract"], feed_subcription_stats
+        )
+
+        feed_item.update(subscription_result)
+
+        new_feed_data.append(feed_item)
+
+    columns = get_feed_column_ids(new_feed_data[0])
+
+    formatted_data = format_table(new_feed_data, columns)
+
+    return columns, formatted_data, new_feed_data
